Added authorization for Users
Now in order to create or change a blog
you must be logged in, and have the correct credentials
This is to prevent other users from changing or deleting posts that
are not their own.
The page build still needs work, and another couple of errors I would
like to figure out.